% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_calc.R
\name{BIC_calc}
\alias{BIC_calc}
\title{Compute Overall BIC for Canonical Directions}
\usage{
BIC_calc(X, Y, U, V)
}
\arguments{
\item{X}{A numeric matrix of predictor variables (n x p).}

\item{Y}{A numeric matrix of response variables (n x m).}

\item{U}{A numeric matrix of canonical weights for X (p x m).}

\item{V}{A numeric matrix of canonical weights for Y (q x m).}
}
\value{
A single numeric value representing the overall BIC across all components.
}
\description{
This function calculates the Bayesian Information Criterion (BIC) for the entire set of
canonical components, based on the Gaussian likelihood between `XU[,j]` and `YV[,j]`.
The number of nonzero entries is determined using a threshold of 10% of the absolute quantiles.
}
