% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVRtesting.R
\name{CVR_testing}
\alias{CVR_testing}
\title{CVR Score Testing for Canonical Direction Weights}
\usage{
CVR_testing(U, X, z, lambda1 = NULL, lambda2 = NULL)
}
\arguments{
\item{U}{A numeric matrix of canonical weights (p x m).}

\item{X}{A numeric matrix of predictor variables (n x p).}

\item{z}{A numeric response vector (n x 1).}

\item{lambda1}{A numeric value for Lasso penalty in regression. If `NULL`, it is determined using cross-validation.}

\item{lambda2}{A numeric value for thresholding. If `NULL`, it is  determined by our procedure.}
}
\value{
A numeric vector of test T-statistics for each canonical component.
}
\description{
Conducts a statistical test to assess the significance of extracted canonical components in characterizing some responses related to mental health.
}
