% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{Ltrinv}
\alias{Ltrinv}
\title{Helper Functions for Low-Rank CCA and Score Testing}
\usage{
Ltrinv(x, V, d = TRUE, d_vec = 1)
}
\arguments{
\item{x}{A numeric vector representing the upper triangular part of a matrix.}

\item{V}{An integer specifying the dimension of the square matrix.}

\item{d}{Logical, if `TRUE`, includes diagonal elements.}

\item{d_vec}{A numeric value for diagonal scaling (default = 1).}
}
\value{
A symmetric matrix of dimension (V x V).
}
\description{
These functions include matrix transformations and the SCAD penalty function.
Transform Vector into a Symmetric Matrix
}
