% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conn.R
\name{plot_conn}
\alias{plot_conn}
\title{Plot Brain Connectivity Matrix}
\usage{
plot_conn(mat)
}
\arguments{
\item{mat}{A square connectivity matrix.}
}
\value{
A ggplot2 object displaying the heatmap.
}
\description{
This function plots a normalized connectivity matrix with network annotations.
}
