% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{SCAD_func}
\alias{SCAD_func}
\title{Smoothly Clipped Absolute Deviation (SCAD) Penalty Function}
\usage{
SCAD_func(yvpen, lambda_ch = 0.01, gamma = 3)
}
\arguments{
\item{yvpen}{A numeric vector of values to be penalized.}

\item{lambda_ch}{A numeric value for the SCAD penalty tuning parameter (default = 0.01).}

\item{gamma}{A numeric value for SCAD adjustment (default = 3). Must be greater than 2.}
}
\value{
A numeric vector with SCAD-penalized values.
}
\description{
Smoothly Clipped Absolute Deviation (SCAD) Penalty Function
}
